Split off TLS portion of registry.TLS
This enables spoofing of a domain for things that aren't pkg/registry,
e.g. a fake implementation of GCR.
